Concentrate all additions for SparseMatrixCSC wrt. main int one file - This is now SparseCSCInterface.jl - Other changes only in SpkSparseSolver.jl: p::Union{Problem,SparseMatrixCSC}
